Initial work for #1
This changes how Service loads and finds the requests. Kernel::load was
a good idea after all. The only drawback is that I have to change
define_request to Aviator.define_request to avoid polluting the global
namespace with the Aviator DSL. That's not a big deal though compared
to the benefits it brings. Running the Service tests now allows
SimpleCov to detect the request definition code. More to come in the
next few commits.