Relax runtime dependency requirements
It's never a good idea for a gem to have strict dependencies. A
problem with this though is that we will have to be diligent with
every new release of dependencies. There's no surefire way to prevent
breakage but relying on automated tests and the CI helps.

As part of this requirements relaxation, a change was done on the
VCR setup script such that it doesn't care what version of Faraday
is used.